+++ lib/src/widgets/index.dart
@@ -42,3 +42,5 @@ export 'marquee.dart';
 export './chatbot/chatbot.dart';
 export 'gt_search_dialog.dart';
 export 'gt_app_button.dart';
+export './pdf_widget/pdf_widget.dart';
+export 'dropdown_search.dart';